pyproject.toml
[project]
name = "pcdc"
version = "0.1.0"
description = "Predictive Coding Digital Circuit — stateful dynamical PC network in PyTorch"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
gguf = [
    "llama-cpp-python>=0.2.50",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pcdc"]

[project.scripts]
pcdc-mnist = "pcdc.training.train_mnist_pc:main"
pcdc-continual = "pcdc.training.continual_runner:main"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
